+++ packages/react/src/index.tsx
@@ -1,3 +1,3 @@
-export * from './events'
 export * from './refs'
+export * from './server'
 export * from './template'
